Python 3 fixup.

+++ idstools/util.py
@@ -42,5 +42,5 @@ def decode_inet_addr(addr):
     if len(addr) == 4:
         return socket.inet_ntoa(addr)
     else:
-        parts = struct.unpack(">" + "H" * (len(addr) / 2), addr)
+        parts = struct.unpack(">" + "H" * int(len(addr) / 2), addr)
         return ":".join("%04x" % p for p in parts)
